Add clean banner for WebAssembly port
- Add custom version header generator to remove git hash from REPL banner
- Override version header generation in WebAssembly Makefile
- Define MICROPY_BANNER_MACHINE for WebAssembly port
- Clean banner shows '10.0.0-beta.2-24' instead of '10.0.0-beta.2-24-ga25f03345a-dirty'

+++ ports/webassembly/objjsproxy.c
@@ -221,7 +221,7 @@ static mp_obj_t jsproxy_call(mp_obj_t self_in, size_t n_args, size_t n_kw, const
         mp_arg_check_num(n_args, n_kw, 0, 1, true);
         uint32_t key[n_kw];
         uint32_t value[PVN * n_kw];
-        for (int i = 0; i < n_kw; ++i) {
+        for (size_t i = 0; i < n_kw; ++i) {
             key[i] = (uintptr_t)mp_obj_str_get_str(args[n_args + i * 2]);
             proxy_convert_mp_to_js_obj_cside(args[n_args + i * 2 + 1], &value[i * PVN]);
         }
@@ -257,7 +257,7 @@ static mp_obj_t jsproxy_call(mp_obj_t self_in, size_t n_args, size_t n_kw, const
         return proxy_convert_js_to_mp_obj_cside(out);
     } else {
         uint32_t value[PVN * n_args];
-        for (int i = 0; i < n_args; ++i) {
+        for (size_t i = 0; i < n_args; ++i) {
             proxy_convert_mp_to_js_obj_cside(args[i], &value[i * PVN]);
         }
         uint32_t out[3];
